+++ tests/testthat/test_index_mome.R
@@ -15,7 +15,8 @@ y ~ b * m1 + cp * x
 indmome := d1 * b
 "
 fitmo1 <- sem(modmo1, dat, meanstructure = TRUE, fixed.x = FALSE)
-fitmo1_boot_out <- do_boot(fitmo1, R = 50, seed = 1234, parallel = FALSE)
+fitmo1_boot_out <- do_boot(fitmo1, R = 50, seed = 1234, parallel = FALSE, progress = FALSE)
+fitmo1_mc_out <- do_mc(fitmo1, R = 200, seed = 1234)
 
 modmo2 <-
 "
@@ -24,7 +25,8 @@ y ~ b * m1 + g4 * w4 + d4 * m1w4 + cp * x
 indmome := d4 * a
 "
 fitmo2 <- sem(modmo2, dat, meanstructure = TRUE, fixed.x = FALSE)
-fitmo2_boot_out <- do_boot(fitmo2, R = 50, seed = 1234, parallel = FALSE)
+fitmo2_boot_out <- do_boot(fitmo2, R = 50, seed = 1234, parallel = FALSE, progress = FALSE)
+fitmo2_mc_out <- do_mc(fitmo2, R = 200, seed = 1234)
 
 modmomo <-
 "
@@ -33,7 +35,8 @@ y ~ b * m1 + g4 * w4 + d4 * m1w4 + cp *x
 indmome := d1 * d4
 "
 fitmomo <- sem(modmomo, dat, meanstructure = TRUE, fixed.x = FALSE)
-fitmomo_boot_out <- do_boot(fitmomo, R = 50, seed = 1234, parallel = FALSE)
+fitmomo_boot_out <- do_boot(fitmomo, R = 50, seed = 1234, parallel = FALSE, progress = FALSE)
+fitmomo_mc_out <- do_mc(fitmomo, R = 200, seed = 1234)
 
 # Adapt to a change in lavaan 0.6-13
 if (packageVersion("lavaan") > "0.6.12") {
@@ -61,6 +64,13 @@ ind_mome2 <- index_of_mome(x = "x", y = "y", m = "m1", w = "w4",
                            fit = fitmo2, boot_ci = TRUE, boot_out = fitmo2_boot_out)
 ind_momome <- index_of_momome(x = "x", y = "y", m = "m1", w = "w1", z = "w4",
                            fit = fitmomo, boot_ci = TRUE, boot_out = fitmomo_boot_out)
+
+ind_mome1_mc <- index_of_mome(x = "x", y = "y", m = "m1", w = "w1",
+                           fit = fitmo1, mc_ci = TRUE, mc_out = fitmo1_mc_out)
+ind_mome2_mc <- index_of_mome(x = "x", y = "y", m = "m1", w = "w4",
+                           fit = fitmo2, mc_ci = TRUE, mc_out = fitmo2_mc_out)
+ind_momome_mc <- index_of_momome(x = "x", y = "y", m = "m1", w = "w1", z = "w4",
+                           fit = fitmomo, mc_ci = TRUE, mc_out = fitmomo_mc_out)
 
 # coef(ind_mome1)
 # confint(ind_mome1)
@@ -106,6 +116,42 @@ test_that("index_of_mome and index_of_momome", {
         unlist(confint(ind_momome)),
         unlist(parameterEstimates(fitmomo_boot)[32, c("ci.lower", "ci.upper")]),
         tolerance = 1e-4,
+        ignore_attr = TRUE
+      )
+  })
+
+test_that("index_of_mome and index_of_momome, Monte Carlo", {
+    expect_equal(
+        coef(ind_mome1_mc),
+        parameterEstimates(fitmo1_boot)[19, "est"],
+        ignore_attr = TRUE
+      )
+    expect_equal(
+        coef(ind_mome2_mc),
+        parameterEstimates(fitmo2_boot)[19, "est"],
+        ignore_attr = TRUE
+      )
+    expect_equal(
+        coef(ind_momome_mc),
+        parameterEstimates(fitmomo_boot)[32, "est"],
+        ignore_attr = TRUE
+      )
+    expect_equal(
+        unlist(confint(ind_mome1_mc)),
+        quantile(ind_mome1_mc$mc_diff, probs = c(.025, .975)),
+        tolerance = 1e-2,
+        ignore_attr = TRUE
+      )
+    expect_equal(
+        unlist(confint(ind_mome2_mc)),
+        quantile(ind_mome2_mc$mc_diff, probs = c(.025, .975)),
+        tolerance = 1e-1,
+        ignore_attr = TRUE
+      )
+    expect_equal(
+        unlist(confint(ind_momome_mc)),
+        quantile(ind_momome_mc$mc_diff, probs = c(.025, .975)),
+        tolerance = 1e-1,
         ignore_attr = TRUE
       )
   })