+++ NEWS.md
@@ -1,7 +1,9 @@
-# manymome 0.1.9.2
+# manymome 0.1.9.5
 
 - Updated badges in README.md. (0.1.9.1)
 - Updated pkgdown site. (0.1.9.2)
+- Added support for Monte Carlo confidence intervals. (0.1.9.3 to 0.1.9.4)
+- Updated some vignettes for Monte Carlo confidence intervals. (0.1.9.5)
 
 # manymome 0.1.9